+++ lib/src/report/dependencies.dart
@@ -224,13 +224,37 @@ Future<ReportSection> trustworthyDependency(PackageContext context) async {
     );
   }
 
+  Future<Subsection> downgrade() async {
+    final message = await context.downgradeAnalysisErrorMessage;
+    final isPassed = message == null;
+    final issues = isPassed
+        ? [
+            RawParagraph(
+                '`pub downgrade` does not expose any static analysis error.'),
+          ]
+        : [
+            Issue(
+                '`pub downgrade` finds static analysis issue(s):\n\n$message'),
+            RawParagraph(
+                'You may run `dart pub upgrade --tighten` to update your dependency constraints.'),
+          ];
+    return Subsection(
+      'Dependency constraint lower bounds are not breaking',
+      issues,
+      isPassed ? 20 : 0,
+      20,
+      isPassed ? ReportStatus.passed : ReportStatus.failed,
+    );
+  }
+
   final dependencySection = await dependencies();
   final sdkSection = await sdkSupport();
-  final subsections = [dependencySection, sdkSection];
+  final downgradeSection = await downgrade();
+  final subsections = [dependencySection, sdkSection, downgradeSection];
   return makeSection(
     id: ReportSectionId.dependency,
     title: 'Support up-to-date dependencies',
-    maxPoints: 20,
+    maxPoints: subsections.map((e) => e.maxPoints).fold(0, (a, b) => a + b),
     subsections: subsections,
     basePath: packageDir,
   );